requirements.txt
# PhotStration Project Requirements
# Professional Photo to Artistic Illustration Renderer

# Core Image Processing
opencv-python>=4.8.0        # Computer vision and image processing
numpy>=1.24.0               # Numerical operations and arrays
Pillow>=10.0.0              # Image manipulation and enhancement
scikit-image>=0.21.0        # Additional image processing algorithms

# Optional Dependencies
matplotlib>=3.7.0           # Plotting and visualization (optional)
scipy>=1.11.0              # Scientific computing (for advanced filters)

# Development Dependencies (Optional)
black>=23.0.0              # Code formatting
pylint>=2.17.0             # Code linting
mypy>=1.5.0               # Type checking

# Additional Utilities
psd-tools>=1.9.0          # PSD file creation and manipulation (optional)